fix: reject null metadata filter values in cluster variable search
A request like {"metadata": {"kind": null}} deserializes to a null filter
object, which reached mapToUntypedOperations() -> filter.getClass() and
threw an NPE, surfacing as a 500 instead of a validation error. Null
values now produce a validation error and are skipped, returning a 400.
